refactor: clean up unused includes and improve code formatting in various files

+++ src/ScoreManager.cpp
@@ -14,7 +14,7 @@ void ScoreManager::UpdateScore() {
         const auto digit = score % 10;
         score /= 10;
         auto scoreObj = std::make_shared<UIObject>(RESOURCE_DIR"/GUI/" + std::to_string(digit) + ".png");
-        scoreObj->SetPosition({m_Offset.x - 40.0f * i, m_Offset.y});
+        scoreObj->SetPosition({m_Offset.x - 40.0f * static_cast<float>(i), m_Offset.y});
         scoreObj->SetScale(0.8f);
         scoreObj->SetZIndex(20); //in game UI:10, Physics2D: 4, slingshot: 3 & 5, popMenu 20, score 20
         m_Scores.push_back(scoreObj);
